requirement.txt
certifi==2019.11.28
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
Pillow==7.0.0
six==1.13.0
torch==1.4.0
torchvision==0.5.0
